Refactored the remote config logic out of the profile class. Now it's handled by a new class called RemoteConfig and is cached in a file called "cached-remote-config.json".
